fix: keep dist after legacy has built

+++ scripts/build-and-package.sh
@@ -13,11 +13,17 @@ case "$1" in
 "extension") npm run make-extension ;;
 
 *)
-    echo "Unknown command: $1"
-    exit 1
-    ;;
+  echo "Unknown command: $1"
+  exit 1
+  ;;
 esac
 
 mv dist "${DIST_NAME}"
 zip -rJ9 "${DIST_NAME}.zip" "${DIST_NAME}"
-rm -rf "${DIST_NAME}"
+
+if [ "$1" == "legacy" ]; then
+  # For upcoming extension build
+  mv "${DIST_NAME}" dist
+else
+  rm -rf "${DIST_NAME}"
+fi
